Using int instead of short for IRotatable.

+++ src/main/java/org/halvors/electrometrics/common/tileentity/IRotatable.java
@@ -1,14 +1,14 @@
 package org.halvors.electrometrics.common.tileentity;
 
 /**
- * This makes a TileEntity rotable, it's meant to be extended.
+ * This makes a TileEntity rotatable, it's meant to be extended.
  *
  * @author halvors
  */
 public interface IRotatable {
 	boolean canSetFacing(int facing);
 
-	short getFacing();
+	int getFacing();
 
-	void setFacing(short facing);
+	void setFacing(int facing);
 }
